requirements.txt
Django>=5.0
psycopg2-binary>=2.9
python-dotenv
drf-yasg==1.21.7
djangorestframework
setuptools